Bug 505520: New nextSlave functions: _nextFastReservedSlave, _nextSlowIdleSlave. r=bhearsum
_nextFastReservedSlave supports having a certain number of fast slaves
'reserved' for a particular purpose (the intent is for release builders to use
these reserved slaves).  Use the new function, 'watchReservedFile' to update
the number of reserved slaves based on the contents of a file that is checked
every minute.

_nextSlowIdleSlave(N) ensures that at least N slow slaves are idle before
starting to assign jobs to slaves.

--HG--
branch : buildbot-0.8.0
extra : rebase_source : edc826109565504aac56b81470e00f1e185e9290